# swagger_devnet.yaml
openapi: 3.0.1
info:
  title: Bitmap Community Marketplace API
  description: "Backend implementation marketplace api endpoints focked ordinals bot marketplace api."
  version: 0.0.1
servers:
  - url: "https://ordinals-marketplace-api-backend.onrender.com/api"
  - url: "http://localhost:9000/api"
...
paths:
...
  /relist:
    post:
      summary: This endpoint updates an existing marketplace listing with a new price.
      description: This endpoint updates an existing marketplace listing with a new price. Calling the endpoint returns a psbt which needs to be signed by the seller's ordinal wallet address. Once signed the signed psbt needs to be returned by calling /marketplace/confirm-relist. if confirm-relist is not called then the existing ordinal listing remains unchanged.
      tags:
        - Listing
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateListing"
      responses:
        "200":
          description: Update Listing Request Success!
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Update Listing Request failed!
          content:
            application/json:
              schema:
                type: string
  /confirm-relist:
    post:
      summary: This endpoint updates an existing marketplace listing with a new price.
      description: This endpoint updates an existing marketplace listing with a new price. Calling the endpoint returns a psbt which needs to be signed by the seller's ordinal wallet address. Once signed the signed psbt needs to be returned by calling /marketplace/confirm-relist. if confirm-relist is not called then the existing ordinal listing remains unchanged.
      tags:
        - Listing
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmUpdateListing"
      responses:
        "200":
          description: Delete Listing Success!
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Delete Listing failed!
          content:
            application/json:
              schema:
                type: string
...
components:
  schemas:
...
    UpdateListing:
      type: object
      properties:
        sellerOrdinalId:
          type: string
          example: bfd7f4c27a4d717e3e78445ceed002a564fcdc9179282d3b515649c461e6a670i0
        sellerOrdinalPrice:
          type: string
          example: 100000
        sellerPaymentAddress:
          type: string
          example: tb1p0sd5xq6sz0eg3r9j5df0qk38pgnuqreav2qqtq5jfvwpk3yhzuxqjyttjy
        sellerOrdinalPublicKey:
          type: string
          example: 03cde4d7fa3f66b13c61279a3a78fd3623428bc69d7e65c770a0fdfd6ea3b0758d
      required:
        - sellerOrdinalId
        - sellerOrdinalPrice
        - sellerPaymentAddress
        - sellerOrdinalPublicKey
    ConfirmUpdateListing:
      type: object
      properties:
        sellerOrdinalId:
          type: string
          example: bfd7f4c27a4d717e3e78445ceed002a564fcdc9179282d3b515649c461e6a670i0
        sellerOrdinalPrice:
          type: string
          example: 200000
        sellerPaymentAddress:
          type: string
          example: tb1p0sd5xq6sz0eg3r9j5df0qk38pgnuqreav2qqtq5jfvwpk3yhzuxqjyttjy
        sellerOrdinalPublicKey:
          type: string
          example: 03cde4d7fa3f66b13c61279a3a78fd3623428bc69d7e65c770a0fdfd6ea3b0758d
        signedListingPSBT:
          type: string
          example: 70736274ff01005e020000000170a6e661c44956513b2d287991dcfc64a502d0ee5c44783e7e714d7ac2f4d7bf0100000000ffffffff01a0860100000000002251207c1b43035013f2888cb2a352f05a270a27c00f3d62800582924b1c1b4497170c000000000001012b22020000000000002251207c1b43035013f2888cb2a352f05a270a27c00f3d62800582924b1c1b4497170c0113408a2e733456fc69d66ca12a17807a3037416fc2debd7156906c08c74ae500e4c412dd9f7fedca14dbf07b0bcc8cf00b24dc66a22921461a18f1d41a03b369de2c011720cde4d7fa3f66b13c61279a3a78fd3623428bc69d7e65c770a0fdfd6ea3b0758d0000
      required:
        - sellerOrdinalId
        - sellerOrdinalPrice
        - sellerPaymentAddress
        - sellerOrdinalPublicKey
        - signedListingPSBT
...